Add Scourge Invasion SQL and update managers
Add full Scourge Invasion SQL data and a hotfix for the 3.3.5 world event. Two new SQL updates were added (full dataset and a hotfix) that include roughly 1283 creature spawns and ~1306 gameobject entries. Also update GameEventMgr.cpp and ObjectMgr.cpp to handle the new invasion spawn/event data and related loading/handling adjustments.

+++ src/server/game/Events/GameEventMgr.cpp
@@ -389,7 +389,7 @@ void GameEventMgr::LoadFromDB()
                 Field* fields = result->Fetch();
 
                 ObjectGuid::LowType guid = fields[0].GetUInt32();
-                int16 event_id = fields[1].GetInt8();
+                int16 event_id = fields[1].GetInt16();
 
                 int32 internal_event_id = mGameEvent.size() + event_id - 1;
 
@@ -439,7 +439,7 @@ void GameEventMgr::LoadFromDB()
                 Field* fields = result->Fetch();
 
                 ObjectGuid::LowType guid = fields[0].GetUInt32();
-                int16 event_id = fields[1].GetInt8();
+                int16 event_id = fields[1].GetInt16();
 
                 int32 internal_event_id = mGameEvent.size() + event_id - 1;
 
@@ -912,7 +912,7 @@ void GameEventMgr::LoadFromDB()
                 Field* fields = result->Fetch();
 
                 uint32 entry   = fields[0].GetUInt32();
-                int16 event_id = fields[1].GetInt8();
+                int16 event_id = fields[1].GetInt16();
 
                 int32 internal_event_id = mGameEvent.size() + event_id - 1;
